+++ src/main/java/org/jruby/rack/UnmappedRackFilter.java
@@ -13,11 +13,11 @@ import java.util.Collection;
 import java.util.Collections;
 import java.util.Set;
 import java.util.stream.Collectors;
-import javax.servlet.FilterChain;
-import javax.servlet.FilterConfig;
-import javax.servlet.ServletException;
-import javax.servlet.ServletResponse;
 
+import jakarta.servlet.FilterChain;
+import jakarta.servlet.FilterConfig;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.ServletResponse;
 import org.jruby.rack.servlet.RequestCapture;
 import org.jruby.rack.servlet.ResponseCapture;