+++ src/main/java/org/jruby/rack/servlet/RewindableInputStream.java
@@ -12,8 +12,9 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.io.RandomAccessFile;
 import java.nio.ByteBuffer;
-import javax.servlet.ReadListener;
-import javax.servlet.ServletInputStream;
+
+import jakarta.servlet.ReadListener;
+import jakarta.servlet.ServletInputStream;
 
 /**
  * Originally inspired by Kirk's RewindableInputStream ...