+++ src/main/java/org/jruby/rack/servlet/ServletRackEnvironment.java
@@ -8,11 +8,11 @@
 package org.jruby.rack.servlet;
 
 import java.io.IOException;
-import javax.servlet.ServletInputStream;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletRequestWrapper;
-import javax.servlet.http.HttpServletResponse;
 
+import jakarta.servlet.ServletInputStream;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletRequestWrapper;
+import jakarta.servlet.http.HttpServletResponse;
 import org.jruby.rack.RackContext;
 import org.jruby.rack.RackEnvironment;