+++ src/main/java/org/jruby/rack/servlet/ServletRackResponseEnvironment.java
@@ -7,9 +7,8 @@
 
 package org.jruby.rack.servlet;
 
-import javax.servlet.http.HttpServletResponse;
-import javax.servlet.http.HttpServletResponseWrapper;
-
+import jakarta.servlet.http.HttpServletResponse;
+import jakarta.servlet.http.HttpServletResponseWrapper;
 import org.jruby.rack.RackResponseEnvironment;
 
 /**