+++ src/main/ruby/rack/handler/servlet/servlet_env.rb
@@ -52,7 +52,7 @@ module Rack
         POST_PARAM_METHODS = [ 'POST', 'PUT', 'DELETE' ].freeze
 
         # Load parameters into the (Rack) env from the Servlet API.
-        # using javax.servlet.http.HttpServletRequest#getParameterMap
+        # using Java::JakartaServletHttp::HttpServletRequest#getParameterMap
         def load_parameters
           get_only = ! POST_PARAM_METHODS.include?( @servlet_env.getMethod )
           # we only need to really do this for POSTs but we'll handle all
@@ -147,7 +147,7 @@ module Rack
         COOKIE_HASH = "rack.request.cookie_hash".freeze
 
         # Load cookies into the (Rack) env from the Servlet API.
-        # using javax.servlet.http.HttpServletRequest#getCookies
+        # using Java::JakartaServletHttp::HttpServletRequest#getCookies
         def load_cookies
           cookie_hash = {}
           (@servlet_env.getCookies || []).each do |cookie|
@@ -178,7 +178,7 @@ module Rack
         end
 
         def parse_query_string
-          Java::JavaxServletHttp::HttpUtils.parseQueryString(query_string)
+          Java::OrgJrubyRackServlet::HttpUtils.parseQueryString(query_string)
         end
 
         def mark_parameter_error(msg)